Initial klayout parallel DRC checker implementation
Works well, with significant performance improvements. This is possible
because one-rule-after-another cannot maintain CPUs fed throughout the
execution. Worse yet, it seems that when run from the gui, only a single
core at most was be used (as opposed to starting it from command line).
The performance is similar to the python approach, but with the
significant advantage of generating a single report file. It should
actually be a little faster, as the general layer calculation are only
done once with the present method.

Some light gymnastics / hacks are required to end up with a single report
and display it through the gui.

This implementation suffers from a few disadvantages:
* As in run_drc.py, unecessary files are created to hold partial drc checks.
  It is not possible to change this without modifying klayout core.
* For reasons that I ignore, forking in the gui thread seem to have
  issues that the command line approach does not have. For gui, I thus
  spawn a new instance of klayout from command line. This has little
  impact on large designs, but small design drc execution time will be
  lower bounded by it
* The klayout gui interface simply hangs while the check is ongoing.
  Before the check would at least provide some amount of GUI updates.
* WaiveDB is not implemented used (can be added)
* A tmp directory is created to hold temporary files. It is currently
  not cleared at the end of execution. (can be added)

Note that run_drc.py is not affected, as it uses the files in
macros/run_drc_main