klayout lvs: Remove comp as a valid connective layer
This suggestion comes from klayoutmatthias. From the
horse's mouth:
https://github.com/KLayout/klayout/issues/2252#issuecomment-3687491898
However, in the other connections comp is not used directly. Instead the derived shapes (ptap, psd etc.) are used and are properly connected to metal 1 through contact. comp is not.

Because of that, single guard rings are extracted as individual nets without being actually connected to something. Those are exposed as pins because they touch guard rings from other cells. The unwanted pins vanish if you take out comp from the connectivity.

There could be debate on whether to connect comp properly to other
layers, or if it should simply be removed, as it is done presently